requirements.txt
pytest>=7.4.0
python-frontmatter>=1.0.0
markdown>=3.4.0
pyyaml>=6.0.1
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0 